Two "default" cases in switch/case structure
PHP 7 does not allow a switch/case structure to have two defaults. This htmlContactList function is only called from a handful of places within Pika, and in none of these places is there a possibility for anything to be passed in besides 'intake', "case_contact', or nothing. And if it is called with no parameter, the function definition sets $mode to 'contacts'. So it does not appear to me that the default: case should even normally come into play. But if it somehow does, I think it makes more sense for it to be on the case: 'contacts' since this corresponds to the default value of the parameter in the function definition.