delete outcomes tied to case when problem changes
There is a behavior in the outcomes module where if outcomes are recorded, then the problem code is changed on the case, and additional outcomes are recorded, that a case can end up with outcomes from multiple problems recorded. This is undesirable and unexpected behavior for at least one organization I've been working with. 

This patch does not require that any database fields be added, but the following line will have to be added to the case-info.html subtemplate:

<input id="prior_problem" name="prior_problem" type="hidden" value="%%[problem]%%">