bug fix to prevent writing non-existent contact ID number to cases.client_id
In certain situations where staff are merging contact cards while also working within the case info tab, it can be possible for an invalid primary client id to exist within the primary client dropdown on the case info tab. It then becomes possible to write this invalid value back to the database. The case then becomes completely inaccesible to staff. 

This patch checks to make sure that the contact ID number being written to cases.client_id actually exists within the contacts table. If it does not, then an empty string '' is written instead to the cases.client_id field.