Handle calls to getNotes passing too few args
The function getNotes is called from three places in the codebase:
1. modules/case-act.php, where it is called with all 5 arguments specified
2. ops/transfer_case.php where it is called with only two arguments,
3. reports/quick_docket/report.php where it is called with only one argument.

This function as written provides default argument values for only 3 out of the 5 arguments. Prior to PHP version 7.1, passing too few arguments in situations like this would raise only a warning. Beginning with 7.1, it throws an error:
https://www.php.net/manual/en/migration71.incompatible.php

Adding default values for arguments 4 and 5 allows the function to be called with any number of arguments.